Swift 3 Migration

+++ OpenStack Summit/CoreSummit/Video.swift
@@ -6,7 +6,7 @@
 //  Copyright © 2016 OpenStack. All rights reserved.
 //
 
-import struct SwiftFoundation.Date
+import struct Foundation.Date
 
 public struct Video: PresentationMaterial {
     
@@ -26,9 +26,9 @@ public struct Video: PresentationMaterial {
     
     public var dataUploaded: Date
     
-    public var order: Int
+    public var order: Int64
     
-    public var views: Int
+    public var views: Int64
     
     public var event: Identifier
 }
